Add files via upload
Code to use for blockchain. There's a lot we need to fix up for this code. We'll eventually need a function that will create a digital signature using the voting machine's private key. The data variable is also a placeholder right now and we'll need to figure out how to integrate the container which will have the list of voter ID and votes.