Added create_new_chain() and append_block funciton
Created 2 new functions that could be used for server and polling station. create_new_chain() should be only used once during initialization of election. Function will create a new chain starting with a genesis block. append_block() is used whenever someone finished voting at the station. It will append a new block to the chain. Note this this is currently just for one station, and once we use multiple stations, we need to implement a broadcast and consensus algorithm before blocks are appended to the chain.