Added File Receiver (#8)
* Added File Receiver and Send File

Added an implementation that piggybacks off your work and uses your standardized input to accept entire files and write them. It accepts input via bytestream as the following:
filename + ' ' + filedata.
This is made possible by the removal of whitespace as standard breaks.

Send File only takes a name of file and loads it up and writes it as bytestream, sending it through standard outgoing messages implementation.